feat: switch to alpine based image
Using distroless was smaller, but the vast majority of the time it
was desirable to be able to get a shell into the malutki container
for doing things like tcpdump so alpine provides a helpful environment
for a small size tradeoff.

At some point we can provide both, but for now this was the most
common need.